project/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-crispy-forms==1.12.0
django-filter==21.1
djangorestframework==3.12.4
Markdown==3.3.4
model-mommy==2.0.0
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.2
